cockpit.js: split Transport code into its own file
Another straight move.
